# buzzbar-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuzzBar Backend API
  version: "1.0.0"
  description: |
    Dev Phase 1 API (stable).

    Stability guarantee:
    - `/api/v1/**` is considered stable after Phase 1.
    - Future changes must be additive only. Breaking changes require `/api/v2`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/admin/inventory/adjust:
    patch:
      tags: [AdminInventory]
      summary: Adjust inventory quantity (admin/employee)
      security: [{ AdminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variantId, delta]
              properties:
                variantId: { type: string }
                delta: { type: integer }
                reason: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/admin/inventory:
    get:
      tags: [AdminInventory]
      summary: List inventory (admin/employee)
      security: [{ AdminBearerAuth: [] }]
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...